FUSETOOLS-2259 - Use latest FIS version available before RC
- handle fuse qualifier and not only redhat ones
