+++ core/features/org.fusesource.ide.core.feature/feature.xml
@@ -171,7 +171,7 @@
          unpack="false"/>
 
    <plugin
-         id="org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007"
+         id="org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025"
          download-size="0"
          install-size="0"
          version="0.0.0"
