+++ core/plugins/org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025/OSGI-INF/camel2.18.1.redhat-000007-management-service.xml
@@ -6,5 +6,5 @@
    <service>
       <provide interface="org.fusesource.ide.camel.model.service.core.ICamelManagerService"/>
    </service>
-   <property name="camel.version" type="String" value="2.18.1.redhat-000007"/>
+   <property name="camel.version" type="String" value="2.18.1.fuse-000025"/>
 </scr:component>
