+++ core/plugins/org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025/OSGI-INF/l10n/bundle.properties
@@ -1,3 +1,3 @@
 #Properties file for org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007
 Bundle-Vendor = JBoss by Red Hat
-Bundle-Name = Fuse Camel Service Plugin (2.18.1.redhat-000007)
+Bundle-Name = Fuse Camel Service Plugin (2.18.1.fuse-000025)
