+++ core/plugins/org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025/build.properties
@@ -1,0 +1,18 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               OSGI-INF/,\
+               catalogs/,\
+               OSGI-INF/l10n/bundle.properties,\
+               libs/camel-blueprint-2.18.1.fuse-000025.jar,\
+               libs/camel-catalog-2.18.1.fuse-000025.jar,\
+               libs/camel-catalog-provider-karaf-2.18.1.fuse-000025.jar,\
+               libs/camel-catalog-provider-springboot-2.18.1.fuse-000025.jar,\
+               libs/camel-core-2.18.1.fuse-000025.jar,\
+               libs/camel-spring-2.18.1.fuse-000025.jar,\
+               libs/log4j-over-slf4j-1.7.21.jar,\
+               libs/slf4j-api-1.7.21.jar,\
+               libs/slf4j-simple-1.7.21.jar
+src.includes = catalogs/
+
