+++ core/plugins/org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025/pom.xml
@@ -12,12 +12,12 @@
 		<relativePath>..</relativePath>
 	</parent>
 
-	<artifactId>org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007</artifactId>
+	<artifactId>org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025</artifactId>
 	<packaging>eclipse-plugin</packaging>
-	<name>JBoss Fuse Tooling :: Core :: Plugins :: Camel 2.18.1.redhat-000007 Service Implementation</name>
+	<name>JBoss Fuse Tooling :: Core :: Plugins :: Camel 2.18.1.fuse-000025 Service Implementation</name>
 
 	<properties>
-		<bundle.camel.version>2.18.1.redhat-000007</bundle.camel.version>
+		<bundle.camel.version>2.18.1.fuse-000025</bundle.camel.version>
 		<bundle.slf4j.version>1.7.21</bundle.slf4j.version>
 	</properties>
 	<build>
