+++ core/plugins/org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025/src/org/fusesource/ide/camel/model/service/internal/CamelService.java
