--- core/plugins/org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007/.classpath
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry exported="true" kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="libs/camel-blueprint-2.18.1.redhat-000007.jar"/>
-	<classpathentry kind="lib" path="libs/camel-catalog-2.18.1.redhat-000007.jar"/>
-	<classpathentry kind="lib" path="libs/camel-catalog-provider-karaf-2.18.1.redhat-000007.jar"/>
-	<classpathentry kind="lib" path="libs/camel-catalog-provider-springboot-2.18.1.redhat-000007.jar"/>
-	<classpathentry kind="lib" path="libs/camel-core-2.18.1.redhat-000007.jar"/>
-	<classpathentry kind="lib" path="libs/camel-spring-2.18.1.redhat-000007.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
