+++ core/plugins/pom.xml
@@ -19,7 +19,7 @@ xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xs
 		<module>org.fusesource.ide.camel.model.service.v2_17_0_redhat_630187</module>
 		<module>org.fusesource.ide.camel.model.service.v2_17_0_redhat_630224</module>
 		<module>org.fusesource.ide.camel.model.service.v2_17_3</module>
-		<module>org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007</module>
+		<module>org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025</module>
 		<module>org.fusesource.ide.camel.validation</module>
 		<module>org.fusesource.ide.cheatsheets</module>
 		<module>org.fusesource.ide.graph</module>
