+++ core/tests/org.fusesource.ide.camel.model.service.core.tests/META-INF/MANIFEST.MF
@@ -8,5 +8,5 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Require-Bundle: org.assertj.core;bundle-version="2.1.0",
  org.junit;bundle-version="4.12.0",
  org.jboss.tools.locus.mockito;bundle-version="1.9.5",
- org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007;bundle-version="9.1.0"
+ org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025;bundle-version="9.1.0"
 Bundle-Vendor: %Bundle-Vendor
