+++ core/tests/org.fusesource.ide.camel.tests.util/META-INF/MANIFEST.MF
@@ -7,7 +7,7 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Require-Bundle: org.fusesource.ide.branding,
  org.junit;bundle-version="4.12.0",
  org.fusesource.ide.camel.model.service.core;bundle-version="9.1.0",
- org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007;bundle-version="9.1.0",
+ org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025;bundle-version="9.1.0",
  org.fusesource.ide.camel.model.service.core.tests.integration,
  org.fusesource.ide.branding,
  org.eclipse.core.runtime;bundle-version="3.11.1",
