+++ core/tests/org.fusesource.ide.camel.validation.tests.integration/META-INF/MANIFEST.MF
@@ -14,7 +14,7 @@ Require-Bundle: org.assertj.core;bundle-version="2.1.0",
  org.fusesource.ide.camel.model.service.core,
  org.eclipse.wst.validation,
  org.eclipse.ui;bundle-version="3.107.0",
- org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007;bundle-version="9.1.0",
+ org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025;bundle-version="9.1.0",
  org.fusesource.ide.camel.model.service.core.tests.integration,
  org.eclipse.emf.common;bundle-version="2.11.0",
  org.eclipse.graphiti;bundle-version="0.13.0",
