+++ editor/plugins/org.fusesource.ide.projecttemplates/plugin.xml
@@ -314,7 +314,7 @@
       	
       	<project-facet-version
             facet="jst.camel"
-            version="2.18.1.redhat-000007">
+            version="2.18.1.fuse-000025">
        	
         	<constraint>
             	<and>
