+++ editor/plugins/org.fusesource.ide.projecttemplates/template-sources/simple/ose/simple-ose-log-springboot/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-#spring.main.sources=com.mycompany
+#spring.main.sources=org.fromForge
 
 logging.config=classpath:logback.xml
