+++ editor/tests/org.fusesource.ide.camel.editor.tests.integration/META-INF/MANIFEST.MF
@@ -23,7 +23,7 @@ Require-Bundle: org.junit;bundle-version="4.12.0",
  org.eclipse.core.databinding;bundle-version="1.5.0",
  org.eclipse.core.databinding.observable,
  org.fusesource.ide.camel.model.service.core;bundle-version="9.1.0",
- org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007;bundle-version="9.1.0",
+ org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025;bundle-version="9.1.0",
  org.fusesource.ide.camel.editor;bundle-version="9.1.0",
  org.fusesource.ide.camel.model.service.core.tests.integration,
  org.fusesource.ide.branding;bundle-version="9.1.0",
