+++ editor/tests/org.fusesource.ide.projecttemplates.tests.integration/META-INF/MANIFEST.MF
@@ -34,7 +34,7 @@ Require-Bundle: org.junit,
  org.fusesource.ide.camel.model.service.v2_17_0_redhat_630187;bundle-version="9.1.0",
  org.fusesource.ide.camel.model.service.v2_17_0_redhat_630224;bundle-version="9.1.0",
  org.fusesource.ide.camel.model.service.v2_17_3;bundle-version="9.1.0",
- org.fusesource.ide.camel.model.service.v2_18_1_redhat_000007;bundle-version="9.1.0",
+ org.fusesource.ide.camel.model.service.v2_18_1_fuse_000025;bundle-version="9.1.0",
  org.fusesource.ide.camel.editor;bundle-version="9.1.0",
  org.fusesource.ide.launcher;bundle-version="9.1.0",
  org.fusesource.ide.launcher.ui,
