+++ editor/tests/org.fusesource.ide.projecttemplates.tests.integration/src/main/java/org/fusesource/ide/projecttemplates/tests/integration/wizards/FuseIntegrationProjectCreatorRunnableForOSESringBootIT.java
@@ -21,7 +21,7 @@ import org.junit.Test;
 public class FuseIntegrationProjectCreatorRunnableForOSESringBootIT extends FuseIntegrationProjectCreatorRunnableIT {
 	
 	public FuseIntegrationProjectCreatorRunnableForOSESringBootIT() {
-		camelVersion = "2.18.1.redhat-000007";
+		camelVersion = "2.18.1.fuse-000025";
 	}
 	
 	@Test
